move server credential handling into sub-package
