+++ src/main/java/org/jenkinsci/plugins/configfiles/maven/security/server/MavenServerIdRequirement.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.plugins.configfiles.maven.security;
+package org.jenkinsci.plugins.configfiles.maven.security.server;
 
 import javax.annotation.CheckForNull;
