+++ src/main/java/org/jenkinsci/plugins/configfiles/maven/security/server/MavenServerIdSpecification.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.plugins.configfiles.maven.security;
+package org.jenkinsci.plugins.configfiles.maven.security.server;
 
 import hudson.Extension;
 import hudson.Util;
