+++ src/main/resources/org/jenkinsci/plugins/configfiles/maven/security/server/ServerCredentialMapping/help-credentialsId.html
